requirements.txt
absl-py==0.7.1
adabound==0.0.5
addict==2.2.1
alembic==1.1.0
astor==0.8.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
cliff==2.15.0
cmd2==0.9.16
colorama==0.4.1
colorlog==4.0.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
flake8==3.7.8
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
imageio==2.5.0
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.0
numpy==1.17.0
opencv-python==4.1.0.25
optuna==0.15.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.5.1
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pipdeptree==0.13.2
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.2
pyperclip==1.7.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.8
stevedore==1.31.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.2.0
torchvision==0.4.0
tornado==6.0.3
tqdm==4.34.0
traitlets==4.3.2
typing==3.7.4.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wrapt==1.11.2
